Add install and update commands
Drop use of python fire; while it is great for quick CLIs it has some
sharp edges that, when they can at all be fixed, forces the interface
to become a worse API.

Many other decisions were made that I cannot bother recalling at the
moment.

Compiling constraints using make and pip-tools in unit tests is done
with --no-header or the python version would show up in the output
making it harder to define the expected output.