Squash deploy job with stable
Publishing if the unstable job does not pass is probably not a good
idea but typically a commit will be tagged on master only after it has
passed all tests so this is not a big problem. Neither is is it a big
problem that the unstable job runs even though it does not affect
anything. Maybe I will come back to this at some point.