Enable multiple, concurrent environments
There are no tests for this use case but the github actions workflow
has been updated to make use of it which gives it some test coverage.
Pilecap is first installed without deps to ensure we do not install
any unpinned packages that we could use accidentally.

Tox is dropped in CI because it is hard to get it to respect version
locks and it would create only one environment per container anyway.
Tox is kept locally because it makes it easy to iterate over python
versions but the normal steps to install deps and the package are
sidestepped so that pilecap can be bootstrapped and used to install
packages into the environment.

In the Makefile `touch $@` is removed because they were annoying;
I would find myself thinking that a check had passed because there was
no error but in reality the check just had not run.

Json is used to print the environment in the header mostly because it
is easy and somewhat human readable.

Canaria domestica is added as a dev dependency to
* ensure the correct dependencies are installed (I would have added it
  as a run or build dependency since these are more error prone but
  then users would not be able to install this package)
* demonstrate that dev requirements do not interact with shared
  constraints.

The update command is renamed compile to make it easier transitioning
between pilecap and pip-tools.

The --output-file option is added as a way to opt out of the custom
naming without reverting to the make workflow. Make can be used to
generate the same constraints but it is tricky to get both the make
file and the annotations in the constraints file both looking nice.

The compilation and gathering code is refactored to better separate
parts that need to deal with the file system from those that,
theoretically, do not. Functions in the compilation module now take
the destination path as an argument to make it easier getting an
overview of what files are created.

The Marker enum initially had the current_value as value but then
I remembered that deserializing command line arguments using the name
of enum variants couples the code to the interface I changed it and
added the property.